Moved settings package
