+++ app/src/main/AndroidManifest.xml
@@ -24,7 +24,7 @@
         </activity>
 
         <activity
-            android:name=".activity.SettingsActivity"
+            android:name=".settings.SettingsActivity"
             android:label="@string/label_settings" />
 
         <activity
