+++ app/src/main/java/com/numero/material_gallery/components/list/ComponentListActivity.kt
@@ -13,7 +13,6 @@ import com.google.android.play.core.appupdate.AppUpdateManagerFactory
 import com.google.android.play.core.install.model.AppUpdateType
 import com.google.android.play.core.install.model.UpdateAvailability
 import com.numero.material_gallery.R
-import com.numero.material_gallery.activity.*
 import com.numero.material_gallery.components.appbar.bottom.BottomAppBarTypeActivity
 import com.numero.material_gallery.components.appbar.top.TopAppBarTypeActivity
 import com.numero.material_gallery.components.DesignComponent
@@ -38,6 +37,7 @@ import com.numero.material_gallery.components.slider.SliderActivity
 import com.numero.material_gallery.components.snackbar.SnackbarActivity
 import com.numero.material_gallery.components.tab.TabActivity
 import com.numero.material_gallery.components.textfield.TextFieldActivity
+import com.numero.material_gallery.settings.SettingsActivity
 import com.numero.material_gallery.view.ListItemAdapter
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.content_main.*
